Add checker interface
